@@ -3,7 +3,7 @@
     "description": "Mantis Bug Tracker",
     "type": "project",
     "require": {
-        "php": ">=5.5.0",
+        "php": "^7.0 || ^8.0",
         "ext-mbstring": "*",
         "ext-json": "*",
         "slim/slim": "^3.0",
@@ -11,11 +11,11 @@
         "vboctor/disposable_email_checker": "^3.0",
         "adodb/adodb-php": "^5.20",
         "phpmailer/phpmailer": "~6.0",
-        "erusev/parsedown": "^1.7.0",
+        "erusev/parsedown": "^1.7.0, <=1.7.3",
         "dapphp/securimage": "dev-mantis"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8"
+        "phpunit/phpunit": "^6.0"
     },
     "repositories": [
         {
@@ -33,7 +33,7 @@
     ],
     "config": {
         "platform": {
-            "php": "5.5.0"
+            "php": "7.0.0"
         }
     }
 }
