@@ -768,7 +768,7 @@ function print_category_option_list( $p_category_id = 0, $p_project_id = null ) 
 			if( count( $t_cat_arr ) == 1 ) {
 				$p_category_id = (int) $t_cat_arr[0]['id'];
 			} else {
-				echo '<option value="0" disabled hidden';
+				echo '<option value="" disabled hidden';
 				check_selected( $p_category_id, 0 );
 				echo '>';
 				echo string_attribute( lang_get( 'select_option' ) );
@@ -1189,30 +1189,6 @@ function print_project_user_list_option_list2( $p_user_id ) {
 }
 
 /**
- * list of projects that a user is in
- * @param integer $p_user_id             An user identifier.
- * @param boolean $p_include_remove_link Whether to display remove link.
- * @return void
- */
-function print_project_user_list( $p_user_id, $p_include_remove_link = true ) {
-	$t_projects = user_get_assigned_projects( $p_user_id );
-
-	foreach( $t_projects as $t_project_id=>$t_project ) {
-		$t_project_name = string_attribute( $t_project['name'] );
-		$t_view_state = $t_project['view_state'];
-		$t_access_level = $t_project['access_level'];
-		$t_access_level = get_enum_element( 'access_levels', $t_access_level );
-		$t_view_state = get_enum_element( 'project_view_state', $t_view_state );
-
-		echo $t_project_name . ' [' . $t_access_level . '] (' . $t_view_state . ')';
-		if( $p_include_remove_link && access_has_project_level( config_get( 'project_user_threshold' ), $t_project_id ) ) {
-			html_button( 'manage_user_proj_delete.php', lang_get( 'remove_link' ), array( 'project_id' => $t_project_id, 'user_id' => $p_user_id ) );
-		}
-		echo '<br />';
-	}
-}
-
-/**
  * List of projects with which the specified field id is linked.
  * For every project, the project name is listed and then the list of custom
  * fields linked in order with their sequence numbers.  The specified field
@@ -1426,12 +1402,21 @@ function print_manage_project_sort_link( $p_page, $p_string, $p_field, $p_dir, $
 
 /**
  * Print a button which presents a standalone form.
+ *
  * If $p_security_token is OFF, the button's form will not contain a security
  * field; this is useful when form does not result in modifications (CSRF is not
  * needed). If otherwise specified (i.e. not null), the parameter must contain
  * a valid security token, previously generated by form_security_token().
  * Use this to avoid performance issues when loading pages having many calls to
  * this function, such as adm_config_report.php.
+ *
+ * The security token field's name is derived from $p_action_page. If it's a
+ * regular page, we use its name without the '.php' extension. In case of a
+ * plugin page (generated by {@see plugin_page()}, we use the value of the
+ * 'page' parameter, replacing '/' by '_'.
+ *
+ * Note: we assume that $p_action_page is valid - no error handling is done.
+ *
  * @param string $p_action_page    The action page.
  * @param string $p_label          The button label.
  * @param array  $p_args_to_post   Associative array of arguments to be posted, with
@@ -1442,23 +1427,34 @@ function print_manage_project_sort_link( $p_page, $p_string, $p_field, $p_dir, $
  * @return void
  */
 function print_form_button( $p_action_page, $p_label, array $p_args_to_post = null, $p_security_token = null, $p_class = '' ) {
-	$t_form_name = explode( '.php', $p_action_page, 2 );
 	# TODO: ensure all uses of print_button supply arguments via $p_args_to_post (POST)
 	# instead of via $p_action_page (GET). Then only add the CSRF form token if
 	# arguments are being sent via the POST method.
-	echo '<form method="post" action="', htmlspecialchars( $p_action_page ), '" class="form-inline inline single-button-form">';
+	echo '<form method="post" action="', htmlspecialchars( $p_action_page ),
+		'" class="form-inline inline single-button-form">';
+
 	if( $p_security_token !== OFF ) {
-		echo form_security_field( $t_form_name[0], $p_security_token );
+		$t_url = parse_url( $p_action_page );
+		$t_page = basename( $t_url['path'] );
+		if( $t_page == 'plugin.php' ) {
+			# Plugin page
+			parse_str( $t_url['query'], $t_param );
+			$t_form_name = 'plugin_' . str_replace( '/', '_', $t_param['page'] );
+		} else {
+			# Regular page
+			list( $t_form_name ) = explode( '.php', $p_action_page, 2 );
+		}
+		echo form_security_field( $t_form_name, $p_security_token );
 	}
-	if( $p_class !== '') {
-		$t_class = $p_class;
-	} else {
-		$t_class = 'btn btn-primary btn-xs btn-white btn-round';
-	}
-	echo '<button type="submit" class="' . $t_class . '">' . $p_label . '</button>';
+
 	if( $p_args_to_post ) {
 		print_hidden_inputs( $p_args_to_post );
 	}
+
+	if( $p_class == '') {
+		$p_class = 'btn btn-primary btn-xs btn-white btn-round';
+	}
+	echo '<button type="submit" class="' . $p_class . '">' . $p_label . '</button>';
 	echo '</form>';
 }
 
