@@ -37,6 +37,8 @@
  * @uses string_api.php
  * @uses user_api.php
  * @uses utility_api.php
+ *
+ * @noinspection PhpUnhandledExceptionInspection
  */
 
 require_once( 'core.php' );
@@ -92,7 +94,6 @@ access_ensure_global_level( $t_user['access_level'] );
 $t_ldap = ( LDAP == config_get_global( 'login_method' ) );
 
 layout_page_header();
-
 layout_page_begin( 'manage_overview_page.php' );
 
 print_manage_menu( 'manage_user_page.php' );
@@ -103,127 +104,179 @@ print_manage_menu( 'manage_user_page.php' );
 
 <!-- USER INFO -->
 <div id="edit-user-div" class="form-container">
-	<form id="edit-user-form" method="post" action="manage_user_update.php">
-		<div class="widget-box widget-color-blue2">
-			<div class="widget-header widget-header-small">
-				<h4 class="widget-title lighter">
-					<?php print_icon( 'fa-user', 'ace-icon' ); ?>
-					<?php echo lang_get('edit_user_title') ?>
-				</h4>
-			</div>
+<form id="edit-user-form" method="post" action="manage_user_update.php">
+	<?php echo form_security_field( 'manage_user_update' ) ?>
+	<input type="hidden" name="user_id" value="<?php echo $t_user['id'] ?>" />
+
+	<div class="widget-box widget-color-blue2">
+		<div class="widget-header widget-header-small">
+			<h4 class="widget-title lighter">
+				<?php print_icon( 'fa-user', 'ace-icon' ); ?>
+				<?php echo lang_get('edit_user_title') ?>
+			</h4>
+		</div>
+
 		<div class="widget-body">
-		<div class="widget-main no-padding">
-		<div class="form-container">
-		<div class="table-responsive">
-		<table class="table table-bordered table-condensed table-striped">
-		<fieldset>
-			<?php echo form_security_field( 'manage_user_update' ) ?>
-			<!-- Title -->
-			<input type="hidden" name="user_id" value="<?php echo $t_user['id'] ?>" />
+			<div class="widget-main no-padding table-responsive">
+				<table class="table table-bordered table-condensed table-striped">
 
-			<!-- Username -->
-			<tr>
-				<td class="category">
-					<?php echo lang_get( 'username_label' ) ?>
-				</td>
-				<td>
-					<input id="edit-username" type="text" class="input-sm" size="32" maxlength="<?php echo DB_FIELD_SIZE_USERNAME;?>" name="username" value="<?php echo string_attribute( $t_user['username'] ) ?>" />
-				</td>
-			</tr>
+					<!-- Username -->
+					<tr>
+						<td class="category width-30">
+							<label for="edit-username">
+								<?php echo lang_get( 'username_label' ) ?>
+							</label>
+						</td>
+						<td>
+							<input id="edit-username" name="username"
+								   type="text" class="input-sm"
+								   size="32" maxlength="<?php echo DB_FIELD_SIZE_USERNAME;?>"
+								   value="<?php echo string_attribute( $t_user['username'] ) ?>"
+							/>
+						</td>
+					</tr>
 
-			<!-- Realname -->
-			<tr><?php
-			if( $t_ldap && ON == config_get_global( 'use_ldap_realname' ) ) {
-				# With LDAP
-				echo '<td class="category">' . lang_get( 'realname_label' ) . '</td>';
-				echo '<td>';
-				echo string_display_line( user_get_realname( $t_user_id ) );
-				echo '</td>';
-			} else {
-				# Without LDAP ?>
-				<td class="category"><?php echo lang_get( 'realname_label' ) ?></td>
-				<td><input id="edit-realname" type="text" class="input-sm" size="32" maxlength="<?php echo DB_FIELD_SIZE_REALNAME;?>" name="realname" value="<?php echo string_attribute( $t_user['realname'] ) ?>" /></td><?php
-			}
-		?>
-			</tr>
-			<!-- Email -->
-			<tr><?php
-			if( $t_ldap && ON == config_get_global( 'use_ldap_email' ) ) {
-				# With LDAP
-				echo '<td class="category">' . lang_get( 'email_label' ) . '</td>';
-				echo '<td>' . string_display_line( user_get_email( $t_user_id ) ) . '</td>';
-			} else {
-				# Without LDAP
-				echo '<td class="category">' . lang_get( 'email_label' ) . '</td>';
-				echo '<td>';
-				print_email_input( 'email', $t_user['email'] );
-				echo '</td>';
-			} ?>
-			</tr>
-			<!-- Access Level -->
-			<tr>
-				<td class="category">
-					<?php echo lang_get( 'access_level_label' ) ?>
-				</td>
-				<td>
-					<select id="edit-access-level" name="access_level" class="input-sm"><?php
-						$t_access_level = $t_user['access_level'];
-						if( !MantisEnum::hasValue( config_get( 'access_levels_enum_string' ), $t_access_level ) ) {
-							$t_access_level = config_get( 'default_new_account_access_level' );
-						}
-						print_project_access_levels_option_list( (int)$t_access_level ); ?>
-					</select>
-				</td>
-			</tr>
-			<!-- Enabled Checkbox -->
-			<tr>
-				<td class="category">
-					<?php echo lang_get( 'enabled_label' ) ?>
-				</td>
-				<td>
-					<label>
-						<input type="checkbox" class="ace" id="edit-enabled" name="enabled" <?php check_checked( (int)$t_user['enabled'], ON ); ?>>
-						<span class="lbl"></span>
-					</label>
-				</td>
-			</tr>
-			<!-- Protected Checkbox -->
-			<tr>
-				<td class="category">
-					<?php echo lang_get( 'protected_label' ) ?>
-				</td>
-				<td>
-					<label>
-						<input type="checkbox" class="ace" id="edit-protected" name="protected" <?php check_checked( (int)$t_user['protected'], ON ); ?>>
-						<span class="lbl"></span>
-					</label>
-				</td>
-			</tr>
+					<!-- Realname -->
+					<tr>
+<?php
+	if( $t_ldap && ON == config_get_global( 'use_ldap_realname' ) ) {
+		# With LDAP
+?>
+						<td class="category">
+							<?php echo lang_get( 'realname_label' ) ?>
+						</td>
+						<td>
+							<?php echo string_display_line( user_get_realname( $t_user_id ) ) ?>
+						</td>
+<?php
+	} else {
+		# Without LDAP
+?>
+						<td class="category">
+							<label for="edit-realname">
+								<?php echo lang_get( 'realname_label' ) ?>
+							</label>
+						</td>
+						<td>
+							<input id="edit-realname" name="realname"
+								   type="text" class="input-sm"
+								   size="32" maxlength="<?php echo DB_FIELD_SIZE_REALNAME;?>"
+								   value="<?php echo string_attribute( $t_user['realname'] ) ?>"
+							/>
+						</td>
+<?php
+	}
+?>
+					</tr>
 
-			<?php event_signal( 'EVENT_MANAGE_USER_UPDATE_FORM', array( $t_user['id'] ) ); ?>
+					<!-- Email -->
+					<tr>
+<?php
+	if( $t_ldap && ON == config_get_global( 'use_ldap_email' ) ) {
+		# With LDAP
+?>
+						<td class="category">
+							<?php echo lang_get( 'email_label' ) ?>
+						</td>
+						<td>
+							<?php echo string_display_line( user_get_email( $t_user_id ) ) ?>
+						</td>
+<?php
+	} else {
+		# Without LDAP
+?>
+						<td class="category">
+							<label for="edit-realname">
+								<?php echo lang_get( 'email_label' ) ?>
+							</label>
+						</td>
+						<td>
+							<?php print_email_input( 'email', $t_user['email'] ); ?>
+						</td>
+<?php
+	}
+?>
+					</tr>
 
-			<!-- Submit Button -->
-		</fieldset>
-		</table>
-		</div>
-		</div>
+					<!-- Access Level -->
+					<tr>
+						<td class="category">
+							<label for="edit-access-level">
+								<?php echo lang_get( 'access_level_label' ) ?>
+							</label>
+						</td>
+						<td>
+							<select id="edit-access-level" name="access_level" class="input-sm">
+<?php
+	$t_access_level = $t_user['access_level'];
+	if( !MantisEnum::hasValue( config_get( 'access_levels_enum_string' ), $t_access_level ) ) {
+		$t_access_level = config_get( 'default_new_account_access_level' );
+	}
+	print_project_access_levels_option_list( (int)$t_access_level );
+?>
+							</select>
+						</td>
+					</tr>
+
+					<!-- Enabled Checkbox -->
+					<tr>
+						<td class="category">
+							<?php echo lang_get( 'enabled_label' ) ?>
+						</td>
+						<td>
+							<label>
+								<input id="edit-enabled" name="enabled"
+									   type="checkbox" class="ace"
+									   <?php check_checked( (int)$t_user['enabled'], ON ); ?>
+								/>
+								<span class="lbl"></span>
+							</label>
+						</td>
+					</tr>
+
+					<!-- Protected Checkbox -->
+					<tr>
+						<td class="category">
+							<?php echo lang_get( 'protected_label' ) ?>
+						</td>
+						<td>
+							<label>
+								<input id="edit-protected" name="protected"
+									   type="checkbox" class="ace"
+									   <?php check_checked( (int)$t_user['protected'], ON ); ?>
+								/>
+								<span class="lbl"></span>
+							</label>
+						</td>
+					</tr>
+
+					<?php event_signal( 'EVENT_MANAGE_USER_UPDATE_FORM', array( $t_user['id'] ) ); ?>
+				</table>
+			</div>
 		</div>
 
 		<div class="widget-toolbox padding-8 clearfix">
-			<input type="submit" class="btn btn-primary btn-white btn-round" value="<?php echo lang_get( 'update_user_button' ) ?>" />
-			<?php
-			if( config_get( 'enable_email_notification' ) == ON ) { ?>
-				&nbsp;
-				<label class="inline">
-					<input type="checkbox" class="ace" id="send-email" name="send_email_notification" checked="checked">
-					<span class="lbl"> <?php echo lang_get( 'notify_user' ) ?></span>
-				</label>
-			<?php } ?>
+			<button class="btn btn-primary btn-white btn-round">
+				<?php echo lang_get( 'update_user_button' ) ?>
+			</button>
+<?php
+	if( config_get( 'enable_email_notification' ) == ON ) {
+?>
+			&nbsp;
+			<label class="inline">
+				<input id="send-email" name="send_email_notification"
+					   type="checkbox" class="ace" checked="checked"
+				/>
+				<span class="lbl">
+					<?php echo lang_get( 'notify_user' ) ?>
+				</span>
+			</label>
+<?php } ?>
 		</div>
-		</div>
-		</div>
-	</form>
+	</div>
+</form>
 </div>
+
 <div class="space-10"></div>
 <?php
 # User action buttons: RESET/UNLOCK and DELETE
@@ -233,7 +286,9 @@ $t_reset = $t_user['id'] != auth_get_current_user_id()
 	&& user_is_enabled( $t_user['id'] )
 	&& !user_is_protected( $t_user['id'] );
 $t_unlock = !user_is_login_request_allowed( $t_user['id'] );
-$t_delete = !( ( user_is_administrator( $t_user_id ) && ( user_count_level( config_get_global( 'admin_site_threshold' ) ) <= 1 ) ) );
+$t_delete = !( user_is_administrator( $t_user_id )
+	&& user_count_level( config_get_global( 'admin_site_threshold' ) ) <= 1
+);
 $t_impersonate = auth_can_impersonate( $t_user['id'] );
 
 if( $t_reset || $t_unlock || $t_delete || $t_impersonate ) {
@@ -245,37 +300,33 @@ if( $t_reset || $t_unlock || $t_delete || $t_impersonate ) {
 <!-- Reset/Unlock Button -->
 <?php if( $t_reset || $t_unlock ) { ?>
 	<form id="manage-user-reset-form" method="post" action="manage_user_reset.php" class="pull-left">
-		<fieldset>
-			<?php echo form_security_field( 'manage_user_reset' ) ?>
-			<input type="hidden" name="user_id" value="<?php echo $t_user['id'] ?>" />
-<?php	if( $t_reset ) { ?>
-			<span><input type="submit" class="btn btn-primary btn-white btn-round" value="<?php echo lang_get( 'reset_password_button' ) ?>" /></span>
-<?php	} else { ?>
-			<span><input type="submit" class="btn btn-primary btn-white btn-round" value="<?php echo lang_get( 'account_unlock_button' ) ?>" /></span>
-<?php	} ?>
-		</fieldset>
+		<?php echo form_security_field( 'manage_user_reset' ) ?>
+		<input type="hidden" name="user_id" value="<?php echo $t_user['id'] ?>" />
+		<button class="btn btn-primary btn-white btn-round">
+			<?php echo lang_get( $t_reset ? 'reset_password_button' : 'account_unlock_button' ) ?>
+		</button>
 	</form>
 <?php } ?>
 
 <!-- Delete Button -->
 <?php if( $t_delete ) { ?>
 	<form id="manage-user-delete-form" method="post" action="manage_user_delete.php" class="pull-left">
-		<fieldset>
-			<?php echo form_security_field( 'manage_user_delete' ) ?>
-			<input type="hidden" name="user_id" value="<?php echo $t_user['id'] ?>" />
-			<span><input type="submit" class="btn btn-primary btn-white btn-round" value="<?php echo lang_get( 'delete_user_button' ) ?>" /></span>
-		</fieldset>
+		<?php echo form_security_field( 'manage_user_delete' ) ?>
+		<input type="hidden" name="user_id" value="<?php echo $t_user['id'] ?>" />
+		<button class="btn btn-primary btn-white btn-round">
+			<?php echo lang_get( 'delete_user_button' ) ?>
+		</button>
 	</form>
 <?php } ?>
 
 <!-- Impersonate Button -->
 <?php if( $t_impersonate ) { ?>
 	<form id="manage-user-impersonate-form" method="post" action="manage_user_impersonate.php" class="pull-left">
-		<fieldset>
-			<?php echo form_security_field( 'manage_user_impersonate' ) ?>
-			<input type="hidden" name="user_id" value="<?php echo $t_user['id'] ?>" />
-			<input type="submit" class="btn btn-primary btn-white btn-round" value="<?php echo lang_get( 'impersonate_user_button' ) ?>" />
-		</fieldset>
+		<?php echo form_security_field( 'manage_user_impersonate' ) ?>
+		<input type="hidden" name="user_id" value="<?php echo $t_user['id'] ?>" />
+		<button class="btn btn-primary btn-white btn-round">
+			<?php echo lang_get( 'impersonate_user_button' ) ?>
+		</button>
 	</form>
 <?php } ?>
 
@@ -290,7 +341,9 @@ if( $t_reset || $t_unlock || $t_delete || $t_impersonate ) {
 <?php
 	print_icon( 'fa-info-circle' );
 	echo '&nbsp;';
-	if( ( ON == config_get( 'send_reset_password' ) ) && ( ON == config_get( 'enable_email_notification' ) ) ) {
+	if( ( ON == config_get( 'send_reset_password' ) )
+		&& ( ON == config_get( 'enable_email_notification' ) )
+	) {
 		echo lang_get( 'reset_password_msg' );
 	} else {
 		echo lang_get( 'reset_password_msg2' );
@@ -304,66 +357,156 @@ if( $t_reset || $t_unlock || $t_delete || $t_impersonate ) {
 
 <div class="clearfix"></div>
 
-<!-- PROJECT ACCESS (if permissions allow) and user is not ADMINISTRATOR -->
-<?php if( access_has_global_level( config_get( 'manage_user_threshold' ) ) &&
-	!user_is_administrator( $t_user_id ) ) {
-?>
-<div class="space-10"></div>
-<div class="widget-box widget-color-blue2">
-<div class="widget-header widget-header-small">
-	<h4 class="widget-title lighter">
-		<?php print_icon( 'fa-puzzle-piece', 'ace-icon' ); ?>
-		<?php echo lang_get('add_user_title') ?>
-	</h4>
-</div>
+<?php
+# Project access sections are only shown if the current user's permissions
+# allow and the user being edited is not an Administrator.
+if( access_has_global_level( config_get( 'manage_user_threshold' ) )
+	&& !user_is_administrator( $t_user_id )
+) {
 
-<div class="widget-body">
-<div class="widget-main no-padding">
-<div class="form-container">
-<div class="table-responsive">
-	<table class="table table-bordered table-condensed table-striped">
-        <tr>
-            <td class="category">
-                <?php echo lang_get( 'assigned_projects_label' ) ?>
-            </td>
-            <td><?php print_project_user_list( $t_user['id'] ) ?></td>
-        </tr>
-        <form id="manage-user-project-add-form" method="post" action="manage_user_proj_add.php">
-        <fieldset>
-            <?php echo form_security_field( 'manage_user_proj_add' ) ?>
-            <input type="hidden" name="user_id" value="<?php echo $t_user['id'] ?>" />
-        <tr>
-            <td class="category">
-                <?php echo lang_get( 'unassigned_projects_label' ) ?>
-            </td>
-            <td>
-                <select id="add-user-project-id" name="project_id[]" class="input-sm" multiple="multiple" size="5">
-                    <?php print_project_user_list_option_list2( $t_user['id'] ) ?>
-                </select>
-            </td>
-        </tr>
-        <tr>
-            <td class="category">
-                <?php echo lang_get( 'access_level_label' ) ?>
-            </td>
-            <td>
-                <select id="add-user-project-access" name="access_level" class="input-sm">
-                    <?php print_project_access_levels_option_list( (int)config_get( 'default_new_account_access_level' ) ) ?>
-                </select>
-            </td>
-        </tr>
-        <tr>
-            <td colspan="2">
-                <input type="submit" class="btn btn-primary btn-sm btn-white btn-round" value="<?php echo lang_get( 'add_user_button' ) ?>" />
-            </td>
-        </tr>
-        </fieldset>
-        </form>
-	</table>
+	$t_projects = user_get_assigned_projects( $t_user['id'] );
+	if( !empty( $t_projects ) ) {
+?>
+<!-- ASSIGNED PROJECTS SECTION -->
+<div class="space-10"></div>
+<div id="project-access-div" class="form-container">
+<form id="project-access-form" method="post" action="manage_user_proj_delete.php">
+
+	<?php echo form_security_field( 'manage_user_proj_delete' ) ?>
+	<input name="user_id" type="hidden" value="<?php echo (int)$t_user['id'] ?>" />
+
+	<div class="widget-box widget-color-blue2">
+		<div class="widget-header widget-header-small">
+			<h4 class="widget-title lighter">
+				<?php print_icon( 'fa-puzzle-piece', 'ace-icon' ); ?>
+				<?php echo lang_get( 'assigned_projects_label' ) ?>
+			</h4>
+		</div>
+		<div class="widget-body">
+			<div class="widget-main no-padding table-responsive">
+				<table class="table table-bordered table-condensed table-striped">
+					<thead>
+						<tr>
+							<th><?php echo lang_get( 'remove_link' ) ?></th>
+							<th><?php echo lang_get( 'project_name' ) ?></th>
+							<th><?php echo lang_get( 'access_level' ) ?></th>
+							<th><?php echo lang_get( 'view_status' ) ?></th>
+						</tr>
+					</thead>
+					<tbody>
+<?php
+		$t_projects = user_get_assigned_projects( $t_user['id'] );
+		foreach( $t_projects as $t_project_id => $t_project ) {
+			$t_can_remove = access_has_project_level( config_get( 'project_user_threshold' ), $t_project_id );
+			$t_project_name = string_attribute( $t_project['name'] );
+			$t_access_level = get_enum_element( 'access_levels', $t_project['access_level'] );
+			$t_view_state = get_enum_element( 'project_view_state', $t_project['view_state'] );
+?>
+						<tr>
+							<td class="center">
+<?php
+			if( $t_can_remove ) {
+?>
+								<!--suppress HtmlFormInputWithoutLabel -->
+								<input name="project_id[]" type="checkbox" class="ace"
+									   value="<?php echo $t_project_id ?>"
+								/>
+								<span class="lbl"></span>
+<?php
+			}
+?>
+							</td>
+							<td><?php echo $t_project_name ?></td>
+							<td><?php echo $t_access_level ?></td>
+							<td><?php echo $t_view_state ?></td>
+						</tr>
+<?php
+		} # foreach project
+?>
+					</tbody>
+				</table>
+			</div>
+		</div>
+
+		<div class="widget-toolbox padding-8 clearfix">
+			<label>
+				<input id="project_id_all" name="project_id_all"
+					   type="checkbox" class="ace check_all"
+				/>
+				<span class="lbl">
+					<?php echo lang_get( 'select_all' ) ?>
+				</span>
+			</label>
+			&nbsp;
+			<button class="btn btn-primary btn-white btn-round">
+				<?php echo lang_get( 'remove_link' ) ?>
+			</button>
+		</div>
+	</div>
+</form>
 </div>
-</div>
-</div>
-</div>
+<?php
+	} # end if user has assigned projects
+?>
+
+<!-- ADD USER TO PROJECT SECTION -->
+<div class="space-10"></div>
+<div id="manage-user-project-add-div" class="form-container">
+<form id="manage-user-project-add-form" method="post" action="manage_user_proj_add.php">
+	<?php echo form_security_field( 'manage_user_proj_add' ) ?>
+	<input type="hidden" name="user_id" value="<?php echo $t_user['id'] ?>" />
+
+	<div class="widget-box widget-color-blue2">
+
+		<div class="widget-header widget-header-small">
+			<h4 class="widget-title lighter">
+				<?php print_icon( 'fa-puzzle-piece', 'ace-icon' ); ?>
+				<?php echo lang_get( 'add_user_title' ) ?>
+			</h4>
+		</div>
+
+		<div class="widget-body">
+			<div class="widget-main no-padding table-responsive">
+				<table class="table table-bordered table-condensed table-striped">
+					<tr>
+						<td class="category width-30">
+							<label for="add-user-project-id">
+								<?php echo lang_get( 'unassigned_projects_label' ) ?>
+							</label>
+						</td>
+						<td>
+							<select id="add-user-project-id" name="project_id[]"
+									class="input-sm" multiple="multiple" size="5">
+								<?php print_project_user_list_option_list2( $t_user['id'] ) ?>
+							</select>
+						</td>
+					</tr>
+					<tr>
+						<td class="category">
+							<label for="add-user-project-access">
+								<?php echo lang_get( 'access_level_label' ) ?>
+							</label>
+						</td>
+						<td>
+							<select id="add-user-project-access" name="access_level"
+									class="input-sm">
+								<?php print_project_access_levels_option_list(
+										(int)config_get( 'default_new_account_access_level' ) )
+								?>
+							</select>
+						</td>
+					</tr>
+				</table>
+			</div>
+		</div>
+
+		<div class="widget-toolbox padding-8 clearfix">
+			<button class="btn btn-primary btn-white btn-round">
+				<?php echo lang_get( 'add_user_button' ) ?>
+			</button>
+		</div>
+	</div>
+</form>
 </div>
 
 <?php
