@@ -51,37 +51,39 @@ form_security_validate( 'manage_user_proj_delete' );
 
 auth_reauthenticate();
 
-$f_project_id = gpc_get_int( 'project_id' );
+$f_projects = gpc_get_int_array( 'project_id' );
 $f_user_id = gpc_get_int( 'user_id' );
 
 user_ensure_exists( $f_user_id );
 
 $t_user = user_get_row( $f_user_id );
 
-access_ensure_project_level( config_get( 'project_user_threshold' ), $f_project_id );
-access_ensure_project_level( $t_user['access_level'], $f_project_id );
+# Make sure user is allowed to perform operation on all selected projects
+$t_project_names = array();
+foreach( $f_projects as $t_project_id ) {
+	access_ensure_project_level( config_get( 'project_user_threshold' ), $t_project_id );
+	access_ensure_project_level( $t_user['access_level'], $t_project_id );
 
-$t_project_name = project_get_name( $f_project_id );
+	$t_project_names[] = project_get_name( $t_project_id );
+}
 
 # Confirm with the user
 helper_ensure_confirmed(
 	sprintf( lang_get( 'remove_user_sure_msg' ),
 		string_attribute( user_get_name( $f_user_id ) ),
-		string_attribute( $t_project_name )
+		string_attribute( implode( '", "', $t_project_names ) )
 	),
 	lang_get( 'remove_user_button' )
 );
 
-project_remove_user( $f_project_id, $f_user_id );
-
+foreach( $f_projects as $t_project_id ) {
+	project_remove_user( $t_project_id, $f_user_id );
+}
 form_security_purge( 'manage_user_proj_delete' );
 
-$t_redirect_url = 'manage_user_edit_page.php?user_id=' .$f_user_id;
+$t_redirect_url = 'manage_user_edit_page.php?user_id=' . $f_user_id;
 
 layout_page_header( null, $t_redirect_url );
-
 layout_page_begin( 'manage_overview_page.php' );
-
 html_operation_successful( $t_redirect_url );
-
 layout_page_end();
